fix(core): tighten conversion length checks and avoid u32 overflow in buf_bgr_to_rgb
- buf_bgr_to_rgb (types.rs): input_size/output_size were computed as
  (width * height * 3) as usize, multiplying in u32 before casting —
  silently wraps on large resolutions in release mode.  Cast each
  operand to usize first: width as usize * height as usize * 3.

- convert_to_rgb RAWRGB arm (frame.rs): had no %3 length guard, unlike
  every sibling arm (convert_to_rgba / convert_to_rgba_buffer /
  convert_to_luma / convert_to_luma_buffer).  Add the same
  !data.len().is_multiple_of(3) check with a matching error message.

- convert_to_luma_buffer MJPEG arm (frame.rs): used dest.len() < luma.len()
  (partial-fill) while every other arm in the same function requires an
  exact dest.len() == luma.len() match.  Tighten to != with a consistent
  "destination buffer size mismatch (expected N, got M)" message.
  Update the two tests that pinned the old oversized-dest success into a
  single test that covers both too-small and too-large as Err.

Skipped F3 (GRAY length guard in convert_to_rgba / convert_to_luma):
chunks_exact_mut.zip is bounds-safe so there is no correctness risk.
Skipped F5 (dead arm removal): no arms were provably unreachable.

Update TODO.md with shipped status.

+++ nokhwa-core/src/types.rs
@@ -1916,8 +1916,8 @@ pub fn buf_bgr_to_rgb(
     let width = resolution.width();
     let height = resolution.height();
 
-    let input_size = (width * height * 3) as usize; // BGR is 3 bytes per pixel
-    let output_size = (width * height * 3) as usize; // RGB is 3 bytes per pixel
+    let input_size = width as usize * height as usize * 3; // BGR is 3 bytes per pixel
+    let output_size = width as usize * height as usize * 3; // RGB is 3 bytes per pixel
 
     if data.len() != input_size {
         return Err(NokhwaError::process_frame(
